refactor(tests): move some definitions
