+++ tests/json_infra/helpers/__init__.py
@@ -1,14 +1,10 @@
 """Helpers to load tests from JSON files."""
 
-from typing import List, Type
-
-from .fixtures import Fixture
+from .fixtures import ALL_FIXTURE_TYPES, Fixture, FixturesFile
 from .load_blockchain_tests import BlockchainTestFixture
 from .load_state_tests import StateTestFixture
 
-ALL_FIXTURE_TYPES: List[Type[Fixture]] = [
-    BlockchainTestFixture,
-    StateTestFixture,
-]
+ALL_FIXTURE_TYPES.append(BlockchainTestFixture)
+ALL_FIXTURE_TYPES.append(StateTestFixture)
 
-__all__ = ["ALL_FIXTURE_TYPES", "Fixture"]
+__all__ = ["ALL_FIXTURE_TYPES", "Fixture", "FixturesFile"]
